OS X: compile util
This is a rollup of commits authored by Adar, JD, and me. These changes allow
most of the util module to be compiled on OS X.

Reviewed-on: http://gerrit.cloudera.org:8080/1186
Tested-by: Mike Percy <mpercy@cloudera.com>

+++ src/kudu/util/hdr_histogram.cc
@@ -100,7 +100,7 @@ void HdrHistogram::Init() {
           kMinValidNumSignificantDigits, kMaxValidNumSignificantDigits);
 
   uint32_t largest_value_with_single_unit_resolution =
-      2 * static_cast<uint32_t>(pow(10, num_significant_digits_));
+      2 * static_cast<uint32_t>(pow(10.0, num_significant_digits_));
 
   // We need to maintain power-of-two sub_bucket_count_ (for clean direct
   // indexing) that is large enough to provide unit resolution to at least
@@ -117,7 +117,7 @@ void HdrHistogram::Init() {
       (sub_bucket_count_magnitude >= 1) ? sub_bucket_count_magnitude - 1 : 0;
 
   // sub_bucket_count_ is approx. 10^num_sig_digits (as a power of 2)
-  sub_bucket_count_ = pow(2, sub_bucket_half_count_magnitude_ + 1);
+  sub_bucket_count_ = pow(2.0, sub_bucket_half_count_magnitude_ + 1);
   sub_bucket_mask_ = sub_bucket_count_ - 1;
   sub_bucket_half_count_ = sub_bucket_count_ / 2;
 
@@ -292,7 +292,7 @@ uint64_t HdrHistogram::ValueAtPercentile(double percentile) const {
   uint64_t count_at_percentile =
     static_cast<uint64_t>(((requested_percentile / 100.0) * count) + 0.5); // Round
   // Make sure we at least reach the first recorded entry
-  count_at_percentile = std::max(count_at_percentile, 1LU);
+  count_at_percentile = std::max(count_at_percentile, static_cast<uint64_t>(1));
 
   uint64_t total_to_current_iJ = 0;
   for (int i = 0; i < bucket_count_; i++) {
@@ -472,9 +472,8 @@ void PercentileIterator::IncrementIterationLevel() {
   percentile_level_to_iterate_from_ = percentile_level_to_iterate_to_;
   // TODO: Can this expression be simplified?
   uint64_t percentile_reporting_ticks = percentile_ticks_per_half_distance_ *
-    static_cast<uint64_t>(pow(2,
-      static_cast<uint64_t>(
-        log(100.0 / (100.0 - (percentile_level_to_iterate_to_))) / log(2)) + 1));
+    static_cast<uint64_t>(pow(2.0,
+          static_cast<int>(log(100.0 / (100.0 - (percentile_level_to_iterate_to_))) / log(2)) + 1));
   percentile_level_to_iterate_to_ += 100.0 / percentile_reporting_ticks;
 }
